+++ src/test/java/janggi/domain/player/PlayerTest.java
@@ -1,8 +1,9 @@
-package janggi.domain;
+package janggi.domain.player;
 
 import static org.assertj.core.api.Assertions.assertThatCode;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import janggi.domain.Dynasty;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
