+++ src/test/java/janggi/domain/player/PlayersTest.java
@@ -1,9 +1,10 @@
-package janggi.domain;
+package janggi.domain.player;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatCode;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
 
+import janggi.domain.Dynasty;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
